Fix and consolidate custom model fields.
